+++ src/index.tsx
@@ -10,10 +10,12 @@ import {
     ShadowStyleIOS,
     StyleProp,
     TransformsStyle,
+    ImageRequireSource,
+    Platform,
     AccessibilityProps,
+    ViewProps,
+    ColorValue,
 } from 'react-native'
-
-const FastImageViewNativeModule = NativeModules.FastImageView
 
 export type ResizeMode = 'contain' | 'cover' | 'stretch' | 'center'
 
@@ -75,12 +77,12 @@ export interface ImageStyle extends FlexStyle, TransformsStyle, ShadowStyleIOS {
     borderTopLeftRadius?: number
     borderTopRightRadius?: number
     overlayColor?: string
-    tintColor?: string
     opacity?: number
 }
 
-export interface FastImageProps extends AccessibilityProps {
-    source: Source | number
+export interface FastImageProps extends AccessibilityProps, ViewProps {
+    source?: Source | ImageRequireSource
+    defaultSource?: ImageRequireSource
     resizeMode?: ResizeMode
     fallback?: boolean
 
@@ -115,7 +117,7 @@ export interface FastImageProps extends AccessibilityProps {
      * If supplied, changes the color of all the non-transparent pixels to the given color.
      */
 
-    tintColor?: number | string
+    tintColor?: ColorValue
 
     /**
      * A unique identifier for this element to be used in UI Automation testing scripts.
@@ -128,8 +130,32 @@ export interface FastImageProps extends AccessibilityProps {
     children?: React.ReactNode
 }
 
+const resolveDefaultSource = (
+    defaultSource?: ImageRequireSource,
+): string | number | null => {
+    if (!defaultSource) {
+        return null
+    }
+    if (Platform.OS === 'android') {
+        // Android receives a URI string, and resolves into a Drawable using RN's methods.
+        const resolved = Image.resolveAssetSource(
+            defaultSource as ImageRequireSource,
+        )
+
+        if (resolved) {
+            return resolved.uri
+        }
+
+        return null
+    }
+    // iOS or other number mapped assets
+    // In iOS the number is passed, and bridged automatically into a UIImage
+    return defaultSource
+}
+
 function FastImageBase({
     source,
+    defaultSource,
     tintColor,
     onLoadStart,
     onProgress,
@@ -153,8 +179,9 @@ function FastImageBase({
             <View style={[styles.imageContainer, style]} ref={forwardedRef}>
                 <Image
                     {...props}
-                    style={StyleSheet.absoluteFill}
+                    style={[StyleSheet.absoluteFill, { tintColor }]}
                     source={resolvedSource}
+                    defaultSource={defaultSource}
                     onLoadStart={onLoadStart}
                     onProgress={onProgress}
                     onLoad={onLoad as any}
@@ -168,6 +195,7 @@ function FastImageBase({
     }
 
     const resolvedSource = Image.resolveAssetSource(source as any)
+    const resolvedDefaultSource = resolveDefaultSource(defaultSource)
 
     return (
         <View style={[styles.imageContainer, style]} ref={forwardedRef}>
@@ -176,6 +204,7 @@ function FastImageBase({
                 tintColor={tintColor}
                 style={StyleSheet.absoluteFill}
                 source={resolvedSource}
+                defaultSource={resolvedDefaultSource}
                 onFastImageLoadStart={onLoadStart}
                 onFastImageProgress={onProgress}
                 onFastImageLoad={onLoad}
@@ -198,12 +227,13 @@ const FastImageComponent: React.ComponentType<FastImageProps> = forwardRef(
 
 FastImageComponent.displayName = 'FastImage'
 
-interface FastImageStaticProperties {
+export interface FastImageStaticProperties {
     resizeMode: typeof resizeMode
     priority: typeof priority
     cacheControl: typeof cacheControl
     preload: (sources: Source[]) => void
-    getCachePath: (source: Source) => Promise<string>
+    clearMemoryCache: () => Promise<void>
+    clearDiskCache: () => Promise<void>
 }
 
 const FastImage: React.ComponentType<FastImageProps> &
@@ -216,7 +246,12 @@ FastImage.cacheControl = cacheControl
 FastImage.priority = priority
 
 FastImage.preload = (sources: Source[]) =>
-    FastImageViewNativeModule.preload(sources)
+    NativeModules.FastImageView.preload(sources)
+
+FastImage.clearMemoryCache = () =>
+    NativeModules.FastImageView.clearMemoryCache()
+
+FastImage.clearDiskCache = () => NativeModules.FastImageView.clearDiskCache()
 
 FastImage.getCachePath = (source: Source) =>
     FastImageViewNativeModule.getCachePath(source)